feat:장소에 대한 정보 및 사진 있으면 뜨우는 기능

+++ src/main/java/kr/co/iei/review/model/dto/ReviewDTO.java
@@ -15,7 +15,7 @@ public class ReviewDTO {
  private String memberNickname;
  private String reviewTitle;
  private String reviewContent;
- private int starRate;
+ private float starRate;
  private String reviewDate;
  private int planNo;
  private int placeId;
